refactor: move optimization functions to separate file
